requirements.txt
# Django Modules
django~=4.2.7
django-cors-headers
django-debug-toolbar
django-ipware
django-model-utils

# Django Rest Modules
djangorestframework
dj-rest-auth

# Other Modules
coreapi
cryptography
gunicorn
requests
passlib
psycopg2
pyopenssl
pytz
validators